FIX - les nombres décimaux fonctionnent de nouveaux!

+++ src/codemdr/objects/CodeMdrNumber.java
@@ -4,6 +4,6 @@ import org.jetbrains.annotations.NotNull;
 
 public abstract class CodeMdrNumber extends CodeMdrObj<Number> {
     protected CodeMdrNumber(@NotNull Number value) {
-        super(value.intValue());
+        super(value);
     }
 }
